Modified injectMouseMove to take floats instead of ints. Fixed onAlphaChanged in Window base class to pass WindowEventArgs with correct source window. Added 'getFormattedLineCount' method to Font class.

+++ include/CEGUISystem.h
@@ -358,7 +358,7 @@ public:
 	\return
 		Nothing.
 	*/
-	void	injectMouseMove(int delta_x, int delta_y);
+	void	injectMouseMove(float delta_x, float delta_y);
 
 
 	/*!
